backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
pydantic[email]
python-multipart
aiosmtplib